integrate with confluent kafka
