+++ server/.env
@@ -1,7 +1,7 @@
 BACKEND_ENV=dev
 NODE_SERVER_PORT=8080
-KAFKA_BOOTSTRAP_SERVER=pkc-l6wr6.europe-west2.gcp.confluent.cloud:9092
+KAFKA_BOOTSTRAP_SERVER="pkc-l6wr6.europe-west2.gcp.confluent.cloud:9092"
 IV=NstLDVVicqT3nC8d
 CRYPT_PASSWORD=3um6cpviSmKiMa5Z
-KAFKA_USERNAME=ZF3KHUOAUGWSDEQD
-KAFKA_PASSWORD=4iENieOiSEEBzDgoTcV9EZODIysPHb5j3UxefWsbVoL6hcsPQzqq0EfjMznjwfFS
+KAFKA_USERNAME="63EAURHMZ32P3ZEB"
+KAFKA_PASSWORD="W00ZW5AQDkOjEgv+G/OdEnqxU7MGVHnMPW7erpys8tZhX2TGZVQYG5nC40AS1S5T"
