+++ server/src/service/auth.service.ts
@@ -10,7 +10,7 @@ import { UserDTO } from './dto/user.dto';
 import { FindManyOptions } from 'typeorm';
 import { decode, encode, generateOtp } from '../utils/helper.service';
 import { generateOtpDTO } from './dto/generate-otp.dto';
-import { OtpStatus, OtpTypes, OtpUsage } from '../utils/enums';
+import { EmailSendOperation, OtpStatus, OtpTypes, OtpUsage } from '../utils/enums';
 import { OtpRepository } from '../repository/otp-repository';
 import { InitiateVerifyPhoneNumberDTO } from './dto/initiate-verify-phone-number.dto';
 import { VerifyPhoneNumberDTO } from './dto/verify-phone-number.dto';
@@ -24,6 +24,10 @@ import { CustomerEntity } from '../domain/customer.entity';
 import { WalletType } from '../domain/enumeration/wallet-type';
 import { WalletStatus } from '../domain/enumeration/wallet-status';
 import { WalletService } from './wallet.service';
+import { PasswordResetDTO } from './dto/password-reset.dto';
+import { SendEmailDTO } from './dto/send-email.dto';
+import { ProducerService } from './producer.service';
+import { SendSmsDTO } from './dto/send-sms.dto';
 
 @Injectable()
 export class AuthService {
@@ -34,7 +38,8 @@ export class AuthService {
         @InjectRepository(OtpRepository) private otpRepository: OtpRepository,
         private userService: UserService,
         private customerService: CustomerService,
-        private walletService: WalletService
+        private walletService: WalletService,
+        private producerService: ProducerService
     ) { }
 
     async login(userLogin: UserLoginDTO): Promise<any> {
@@ -137,7 +142,28 @@ export class AuthService {
         return await this.userService.findAndCount(options);
     }
 
-    async getVerifyPhoneNumberOTP(payload:InitiateVerifyPhoneNumberDTO): Promise<any> {
+    async initResetPassword(payload: PasswordResetDTO ): Promise<any> {
+
+        // perform email authentication
+        const authUser = await this.getAccountByEmail(payload.email);
+        if(!authUser)
+            throw new HttpException(`User with email "${payload.email}" not found`, HttpStatus.BAD_REQUEST);
+
+        // generate terminal password reset link and include it in the email payload data
+        // To-Do
+
+        const emailPaylaod: SendEmailDTO = {
+            data: authUser,
+            email: payload.email,
+            emailType: EmailSendOperation.RESET_PASSWORD,
+        };
+        // use the kafka client to emmit a message to the broker with topic "send-email" with the email emailPaylaod
+        await this.producerService.produce({topic: 'send-email', messages: [{value: JSON.stringify(emailPaylaod)}]});
+
+        return {message: 'Email sent!'};  
+    }
+
+    async initVerifyPhoneNumber(payload:InitiateVerifyPhoneNumberDTO): Promise<any> {
 
         // check if phone number has already been used by another user
         const cutomerFind: CustomerDTO = await this.customerService.findByFields({ where: { phoneNumber: payload.phoneNumber } });
@@ -147,9 +173,7 @@ export class AuthService {
 
         // generate otp for phone number verification
         const otp_generation_options: generateOtpDTO = {size: 4, validityDuration: 90 }
-
         const {otp, timestamp, expiration_time} = await generateOtp(otp_generation_options);
-        console.log(expiration_time);
 
         //Create OTP instance in DB
         const otp_instance = await this.otpRepository.save({
@@ -167,10 +191,20 @@ export class AuthService {
             "type": OtpTypes.PHONE_NUMBER,
             "otp_id": otp_instance.id
         }
-
         // Encrypt the details object
         const otpKey = await encode(JSON.stringify(details))
-        return {otp, otpKey}
+        
+        // compose sms message
+        const smsMessage = `Hello, please enter this code ${otp} to verify your phone number`;
+        // build sms payload for kafka messaging
+        const smsPaylaod: SendSmsDTO = {
+            message: smsMessage,
+            phoneNumber: payload.phoneNumber,
+        };
+        // use the kafka client to emmit a message to the broker with topic "send-sms" with the smsPaylaod
+        await this.producerService.produce({topic: 'send-sms', messages: [{value: JSON.stringify(smsPaylaod)}]});
+        // return the otp key to the client
+        return { otpKey };
     }
 
     async verifyPhoneNumber(payload: VerifyPhoneNumberDTO): Promise<any> {
