--- server/src/transportConfig.ts
@@ -1,15 +1,0 @@
-// import { KafkaOptions, Transport } from '@nestjs/microservices';
-// require('dotenv').config();
-
-// export const TransportConfig: KafkaOptions = {
-//     transport: Transport.KAFKA,
-//     options: {
-//         client: {
-//             brokers: process.env.BROKERS.split(' '),
-//         },
-//         consumer: {
-//             groupId: 'core-service',
-//             allowAutoTopicCreation: true,
-//         },
-//     },
-// };
